requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
fpdf2>=2.7.0
python-docx>=0.8.11
markdown>=3.4.0
Pillow>=9.5.0
plotly>=5.14.0
openpyxl>=3.1.2
xlrd>=2.0.1
watchdog>=3.0.0